+++ test/audits/backlinks.test.js
@@ -18,6 +18,8 @@ import sinonChai from 'sinon-chai';
 import nock from 'nock';
 import { FirefallClient } from '@adobe/spacecat-shared-gpt-client';
 import auditDataMock from '../fixtures/broken-backlinks/audit.json' with { type: 'json' };
+import auditDataSuggestionsMock from '../fixtures/broken-backlinks/auditWithSuggestions.json' with { type: 'json' };
+import rumTraffic from '../fixtures/broken-backlinks/all-traffic.json' with { type: 'json' };
 import { brokenBacklinksAuditRunner, opportunityAndSuggestions, generateSuggestionData } from '../../src/backlinks/handler.js';
 import { MockContextBuilder } from '../shared.js';
 import {
@@ -39,6 +41,8 @@ import {
   brokenBacklinksSuggestions,
   suggestions,
 } from '../fixtures/broken-backlinks/suggestion.js';
+import { organicTraffic } from '../fixtures/broken-backlinks/organic-traffic.js';
+import calculateKpiMetrics from '../../src/backlinks/kpi-metrics.js';
 
 use(sinonChai);
 use(chaiAsPromised);
@@ -65,6 +69,7 @@ describe('Backlinks Tests', function () {
           AHREFS_API_BASE_URL: 'https://ahrefs.com',
           AHREFS_API_KEY: 'ahrefs-api',
           S3_SCRAPER_BUCKET_NAME: 'test-bucket',
+          S3_IMPORTER_BUCKET_NAME: 'test-import-bucket',
         },
         s3Client: {
           send: sandbox.stub(),
@@ -124,6 +129,17 @@ describe('Backlinks Tests', function () {
   });
 
   it('should transform the audit result into an opportunity in the post processor and create a new opportunity', async () => {
+    context.s3Client.send.onCall(0).resolves({
+      Body: {
+        transformToString: sinon.stub().resolves(JSON.stringify(rumTraffic)),
+      },
+    });
+
+    context.s3Client.send.onCall(1).resolves({
+      Body: {
+        transformToString: sinon.stub().resolves(JSON.stringify(organicTraffic(site))),
+      },
+    });
     context.dataAccess.Site.findById = sinon.stub().withArgs('site1').resolves(site);
     context.dataAccess.Opportunity.create.resolves(brokenBacklinksOpportunity);
     brokenBacklinksOpportunity.addSuggestions.resolves(brokenBacklinksSuggestions);
@@ -132,13 +148,18 @@ describe('Backlinks Tests', function () {
 
     ahrefsMock(site.getBaseURL(), auditDataMock.auditResult);
 
-    await opportunityAndSuggestions(auditUrl, auditDataMock, context);
+    await opportunityAndSuggestions(auditUrl, auditDataSuggestionsMock, context, site);
+
+    const kpiDeltas = {
+      projectedTrafficLost: sinon.match.number,
+      projectedTrafficValue: sinon.match.number,
+    };
 
     expect(context.dataAccess.Opportunity.create)
       .to
       .have
       .been
-      .calledOnceWith(opportunityData(auditDataMock.siteId, auditDataMock.id));
+      .calledOnceWith(opportunityData(auditDataMock.siteId, auditDataMock.id, kpiDeltas));
     expect(brokenBacklinksOpportunity.addSuggestions).to.have.been.calledOnceWith(suggestions);
   });
 
@@ -152,7 +173,7 @@ describe('Backlinks Tests', function () {
 
     ahrefsMock(site.getBaseURL(), auditDataMock.auditResult);
 
-    await opportunityAndSuggestions(auditUrl, auditDataMock, context);
+    await opportunityAndSuggestions(auditUrl, auditDataSuggestionsMock, context, site);
 
     expect(context.dataAccess.Opportunity.create).to.not.have.been.called;
     expect(brokenBacklinksOpportunity.setAuditId).to.have.been.calledOnceWith(auditDataMock.id);
@@ -177,7 +198,7 @@ describe('Backlinks Tests', function () {
       .reply(200, auditDataMock.auditResult);
 
     try {
-      await opportunityAndSuggestions(auditUrl, auditDataMock, context);
+      await opportunityAndSuggestions(auditUrl, auditDataSuggestionsMock, context, site);
     } catch (e) {
       expect(e.message).to.equal(errorMessage);
     }
@@ -444,6 +465,62 @@ describe('Backlinks Tests', function () {
         },
       ]);
       expect(context.log.error).to.have.been.calledWith('Batch processing error: Firefall error');
+    });
+  });
+
+  describe('calculateKpiMetrics', () => {
+    const auditData = {
+      auditResult: {
+        brokenBacklinks: [
+          { traffic_domain: 25000001, urlsSuggested: ['https://foo.com/bar/redirect'] },
+          { traffic_domain: 10000001, urlsSuggested: ['https://foo.com/bar/baz/redirect'] },
+          { traffic_domain: 10001, urlsSuggested: ['https://foo.com/qux/redirect'] },
+          { traffic_domain: 100, urlsSuggested: ['https://foo.com/bar/baz/qux/redirect'] },
+        ],
+      },
+    };
+
+    it('should calculate metrics correctly for a single broken backlink', async () => {
+      context.s3Client.send.onCall(0).resolves({
+        Body: {
+          transformToString: sinon.stub().resolves(JSON.stringify(rumTraffic)),
+        },
+      });
+
+      context.s3Client.send.onCall(1).resolves({
+        Body: {
+          transformToString: sinon.stub().resolves(JSON.stringify(organicTraffic(site))),
+        },
+      });
+
+      const result = await calculateKpiMetrics(auditData, context, site);
+      expect(result.projectedTrafficLost).to.equal(26788.645);
+      expect(result.projectedTrafficValue).to.equal(534287.974025892);
+    });
+
+    it('skips URL if no RUM data is available for just individual URLs', async () => {
+      delete rumTraffic[0].earned;
+      context.s3Client.send.onCall(0).resolves({
+        Body: {
+          transformToString: sinon.stub().resolves(JSON.stringify(rumTraffic)),
+        },
+      });
+
+      context.s3Client.send.onCall(1).resolves({
+        Body: {
+          transformToString: sinon.stub().resolves(JSON.stringify(organicTraffic(site))),
+        },
+      });
+
+      const result = await calculateKpiMetrics(auditData, context, site);
+      expect(result.projectedTrafficLost).to.equal(14545.045000000002);
+    });
+
+    it('returns early if there is no RUM traffic data', async () => {
+      context.s3Client.send.onCall(0).resolves(null);
+
+      const result = await calculateKpiMetrics(auditData, context, site);
+      expect(result).to.be.null;
     });
   });
 });